% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticHelper.R
\name{mean_sd}
\alias{mean_sd}
\title{Mittelwert und Standardabweichung}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{Ein numerischer Vektor}
}
\value{
Eine Liste mit Mittelwert und Standardabweichung
}
\description{
Diese Funktion berechnet den Mittelwert und die Standardabweichung eines numerischen Vektors.
}
\examples{
mean_sd(c(1, 2, 3, 4, 5))

}
