% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_var.R
\name{median_var}
\alias{median_var}
\title{Median und Varianz berechnen}
\usage{
median_var(x)
}
\arguments{
\item{x}{Ein numerischer Vektor}
}
\value{
Eine Liste mit Median und Varianz
}
\description{
Diese Funktion berechnet den Median und die Varianz eines numerischen Vektors.
}
\examples{
median_var(c(1, 2, 3, 4, 5))

}
